Move testing util classes into testing package
